+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PATCH version when backwards compatible bug **fixes** are implemented.
 
 ## [Unreleased]
+### Fixed
+- Gson register adapters
 
 ## [2.25.1] - 2026-03-20
 ### Fixed